requirements.txt
numpy
sounddevice
openwakeword
playwright
python-dotenv
